created connection factory to change between sql and api clients

+++ project-1/consoleApp/TicketTrackerApp/Models/Ticket.cs
@@ -7,14 +7,14 @@ public class Ticket
     public string Description { get; set; }
 
     public string? Name { get; set; }
-    public int Amount { get; set; }
+    public decimal Amount { get; set; }
 
     public string Status { get; set; } = "pending";
 
     public int EmployeeId { get; set; }
 
 
-    public Ticket(string description, int amount, int employeeId)
+    public Ticket(string description, decimal amount, int employeeId)
     {
         Description = description;
         Amount = amount;
